+++ binnacle/lib/providers/compass_provider.dart
@@ -7,13 +7,10 @@ import '../models/service_data.dart';
 import '../models/provider_data.dart';
 import '../services/service_wrapper_interface.dart';
 
-import '../repository.dart';
-
-
 class CompassProvider {
   ICompassService _currentService;
 
-  ServiceList _serviceList; 
+  ServiceList _serviceList;
   StreamController<CompassModel> _stream = StreamController();
   StreamController<ServiceData> _activeService = StreamController();
   StreamSubscription _subscription;
@@ -21,51 +18,52 @@ class CompassProvider {
 
   ProviderData _providerData = ProviderData('compass', 'manual');
   StreamController _providerDataStream = StreamController();
-  
 
-  CompassProvider(ServiceList serviceList){
+  CompassProvider(ServiceList serviceList) {
     this._serviceList = serviceList;
     _availableServices.sink.add(_serviceList);
     _providerDataStream.sink.add(this._providerData);
 
-  
     setUpService(this._serviceList.defaultService);
   }
 
   setUpService(ServiceWrapper serviceWrapper) {
-    this._currentService = serviceWrapper.service; 
-    this._activeService.sink.add(serviceWrapper.serviceData); 
+    this._currentService = serviceWrapper.service;
+    this._activeService.sink.add(serviceWrapper.serviceData);
 
-    this._subscription = this._currentService.compassStream.stream.listen((data) => this._stream.add(data));
-    _subscription.onError((error)  {
-      if(_providerData.mode == 'auto') {
-        changeService( _serviceList.nextPriority( serviceWrapper.serviceData ).serviceData);
+    this._subscription = this
+        ._currentService
+        .compassStream
+        .stream
+        .listen((data) => this._stream.add(data));
+    _subscription.onError((error) {
+      if (_providerData.mode == 'auto') {
+        changeService(
+            _serviceList.nextPriority(serviceWrapper.serviceData).serviceData);
       }
     });
   }
 
-  changeService(ServiceData serviceData)  async {
+  changeService(ServiceData serviceData) async {
     await this._currentService.dispose();
     ServiceWrapper serviceWrapper = this._serviceList.service(serviceData);
     setUpService(serviceWrapper);
-
   }
+
   toggleMode(ProviderData providerData) {
     ProviderData temp;
-    if ( providerData.mode == "manual" ) {
+    if (providerData.mode == "manual") {
       temp = ProviderData(_providerData.type, 'auto');
     }
-    if( providerData.mode == "auto" ) {
+    if (providerData.mode == "auto") {
       temp = ProviderData(_providerData.type, 'manual');
     }
     this._providerData = temp;
     this._providerDataStream.sink.add(this._providerData);
-
   }
 
   StreamController<CompassModel> get compass => this._stream;
   StreamController<ServiceData> get activeService => this._activeService;
   StreamController<ServiceList> get availableServices => _availableServices;
   StreamController get providerData => _providerDataStream;
-
-} 
+}
