+++ binnacle/lib/ui/arrow_painter.dart
