+++ binnacle/lib/ui/compass/compass_dial_painter.dart
